[Refactor]메인페이지 코드 분리 커스텀 훅 폴더 생성 폴더에 usestate, useCallback을 useModalManagement라는 이름으로 저장 자동 로그아웃 useEffect와 소셜 로그인 useEffect를 저장 메인페이지 코드에 import
