--- client/src/hooks/customHooks/useModal.tsx
@@ -1,16 +1,0 @@
-// hooks/useModal.tsx
-import { useState, useCallback } from "react";
-
-export const useModal = (initialState: boolean = false) => {
-  const [isOpen, setIsOpen] = useState(initialState);
-
-  const openModal = useCallback(() => {
-    setIsOpen(true);
-  }, []);
-
-  const closeModal = useCallback(() => {
-    setIsOpen(false);
-  }, []);
-
-  return { isOpen, openModal, closeModal };
-};
